[diffusion] refactor: refactor realtime control state and adapters (#27698)
