+++ python/sglang/multimodal_gen/runtime/entrypoints/openai/realtime/adapters/sana_wm_realtime_adapter.py
@@ -2,30 +2,18 @@
 
 from __future__ import annotations
 
-import hashlib
-import os
-import tempfile
 from typing import TYPE_CHECKING, Any
-
-from fastapi import WebSocket
 
 from sglang.multimodal_gen.runtime.entrypoints.openai.protocol import (
     RealtimeEvent,
     RealtimeVideoGenerationsRequest,
 )
 from sglang.multimodal_gen.runtime.entrypoints.openai.realtime.realtime_adapter import (
+    BaseRealtimeModelAdapter,
     RealtimeChunkInputs,
-    RealtimeModelAdapter,
+    build_realtime_sampling_params,
+    save_realtime_first_frame,
 )
-from sglang.multimodal_gen.runtime.entrypoints.openai.realtime.realtime_output_adapter import (
-    RawRGBRealtimeOutputAdapter,
-    RealtimeFrameSendStats,
-)
-from sglang.multimodal_gen.runtime.entrypoints.openai.utils import (
-    build_sampling_params,
-    save_image_to_path,
-)
-from sglang.multimodal_gen.runtime.entrypoints.utils import prepare_request
 from sglang.multimodal_gen.runtime.pipelines_core.stages.model_specific_stages.sana_wm.base import (
     normalize_sana_wm_camera_actions,
     parse_sana_wm_action_string,
@@ -34,10 +22,9 @@ from sglang.multimodal_gen.runtime.pipelines_core.stages.model_specific_stages.s
 from sglang.multimodal_gen.runtime.pipelines_core.stages.model_specific_stages.sana_wm.self_forcing import (
     SanaWMSelfForcingSampler,
 )
-from sglang.multimodal_gen.runtime.realtime.camera_controls import (
+from sglang.multimodal_gen.runtime.realtime.states import (
     RealtimeCameraControlState,
 )
-from sglang.multimodal_gen.runtime.server_args import get_global_server_args
 
 if TYPE_CHECKING:
     from sglang.multimodal_gen.runtime.entrypoints.openai.realtime.generate_session import (
@@ -46,7 +33,6 @@ if TYPE_CHECKING:
     )
     from sglang.multimodal_gen.runtime.pipelines_core.schedule_batch import (
         OutputBatch,
-        Req,
     )
     from sglang.multimodal_gen.runtime.server_args import ServerArgs
 
@@ -77,25 +63,20 @@ class SanaWMRealtimeAdapterState(RealtimeCameraControlState):
         super().clear()
         self.base_condition_inputs.clear()
 
-    def receive_camera_event_payload(
+    def receive_camera_control_event_payload(
         self,
         payload: Any,
         *,
         event_id: int | None,
     ) -> str:
-        return super().receive_camera_event_payload(
+        return super().receive_camera_control_event_payload(
             payload,
             event_id=event_id,
             validate_camera_actions=SanaWMRealtimeAdapter._validate_camera_actions,
         )
 
 
-class SanaWMRealtimeAdapter(RealtimeModelAdapter):
-    name = "sana_wm_realtime"
-
-    def __init__(self):
-        self.output_adapter = RawRGBRealtimeOutputAdapter()
-
+class SanaWMRealtimeAdapter(BaseRealtimeModelAdapter):
     def create_state(self) -> SanaWMRealtimeAdapterState:
         return SanaWMRealtimeAdapterState()
 
@@ -122,9 +103,6 @@ class SanaWMRealtimeAdapter(RealtimeModelAdapter):
         session: GenerateSession,
         request: RealtimeVideoGenerationsRequest,
     ) -> None:
-        if request.first_frame is None:
-            raise ValueError("SANA-WM realtime requires first_frame")
-
         request.size = request.size or SANA_WM_DEFAULT_SIZE
         if request.num_frames is not None:
             request.num_frames = int(request.num_frames)
@@ -157,42 +135,21 @@ class SanaWMRealtimeAdapter(RealtimeModelAdapter):
         if camera_actions is not None and action is not None:
             raise ValueError("pass only one of camera_actions or action")
         if camera_actions is not None:
-            state.receive_camera_event_payload(camera_actions, event_id=None)
+            state.receive_camera_control_event_payload(camera_actions, event_id=None)
         if action is not None:
             if not isinstance(action, str) or not action:
                 raise ValueError("action condition input must be a non-empty string")
-            state.receive_camera_script(
+            state.receive_camera_action_script(
                 parse_sana_wm_action_string(action), event_id=None
             )
         state.base_condition_inputs = condition_inputs
 
-        server_args = get_global_server_args()
-        if server_args.input_save_path is not None:
-            uploads_dir = server_args.input_save_path
-            os.makedirs(uploads_dir, exist_ok=True)
-        else:
-            if session.input_temp_dir is None:
-                session.input_temp_dir = tempfile.mkdtemp(prefix="sglang_input_")
-            uploads_dir = session.input_temp_dir
-
-        if isinstance(
-            request.first_frame, str
-        ) and request.first_frame.lower().startswith(("http://", "https://")):
-            suffix = os.path.splitext(request.first_frame.split("?", 1)[0])[1]
-            digest = hashlib.sha256(request.first_frame.encode("utf-8")).hexdigest()[
-                :16
-            ]
-            target_path = os.path.join(uploads_dir, f"realtime_ref_{digest}{suffix}")
-            if os.path.exists(target_path):
-                request.first_frame = target_path
-                return
-        else:
-            target_path = os.path.join(uploads_dir, f"{session.id}_first_frame")
-        image_path = await save_image_to_path(request.first_frame, target_path)
-        request.first_frame = image_path
-
-    async def wait_for_next_chunk(self, session: GenerateSession) -> None:
-        del session
+        await save_realtime_first_frame(
+            session,
+            request,
+            required_error="SANA-WM realtime requires first_frame",
+            cache_remote_urls=True,
+        )
 
     def ingest_event(
         self,
@@ -201,7 +158,7 @@ class SanaWMRealtimeAdapter(RealtimeModelAdapter):
     ) -> str:
         state = self._state(session)
         if event.kind == "camera_actions":
-            return state.receive_camera_event_payload(
+            return state.receive_camera_control_event_payload(
                 event.payload,
                 event_id=event.event_id,
             )
@@ -209,16 +166,23 @@ class SanaWMRealtimeAdapter(RealtimeModelAdapter):
             if not isinstance(event.payload, str) or not event.payload:
                 raise ValueError("action event payload must be a non-empty string")
             camera_actions = parse_sana_wm_action_string(event.payload)
-            state.receive_camera_script(camera_actions, event_id=event.event_id)
+            state.receive_camera_action_script(camera_actions, event_id=event.event_id)
             return f"kind=action, frames={len(camera_actions)}"
         raise ValueError(f"unsupported event kind: {event.kind}")
 
-    def _sample_chunk_inputs(
+    def sample_chunk_inputs(
         self,
         session: GenerateSession,
+        server_args: ServerArgs,
         chunk: RealtimeChunkContext,
-        action_chunk_size: int,
+        chunk_size: int,
     ) -> RealtimeChunkInputs:
+        action_chunk_size = self._action_chunk_size(
+            session,
+            server_args,
+            chunk,
+            chunk_size,
+        )
         state = self._state(session)
         request = session.request
         if request is None:
@@ -233,9 +197,10 @@ class SanaWMRealtimeAdapter(RealtimeModelAdapter):
             condition_inputs=condition_inputs,
         )
 
-    def _build_sampling_params(
+    def build_sampling_params(
         self,
         session: GenerateSession,
+        server_args: ServerArgs,
         chunk: RealtimeChunkContext,
         chunk_inputs: RealtimeChunkInputs,
         chunk_size: int,
@@ -244,49 +209,22 @@ class SanaWMRealtimeAdapter(RealtimeModelAdapter):
         if request is None:
             raise ValueError("realtime request is not initialized")
 
-        return build_sampling_params(
+        return build_realtime_sampling_params(
             chunk.request_id,
-            prompt=chunk_inputs.prompt,
-            size=request.size,
+            request=request,
+            chunk_inputs=chunk_inputs,
             num_frames=request.num_frames,
-            fps=request.fps,
-            image_path=request.first_frame,
-            output_file_name=chunk.request_id,
-            save_output=False,
-            seed=request.seed,
-            generator_device=request.generator_device,
             num_inference_steps=request.num_inference_steps,
-            guidance_scale=request.guidance_scale,
-            guidance_scale_2=request.guidance_scale_2,
-            negative_prompt=request.negative_prompt,
-            enable_teacache=request.enable_teacache,
-            enable_frame_interpolation=request.enable_frame_interpolation,
-            frame_interpolation_exp=request.frame_interpolation_exp,
-            frame_interpolation_scale=request.frame_interpolation_scale,
-            frame_interpolation_model_path=request.frame_interpolation_model_path,
-            enable_upscaling=request.enable_upscaling,
-            upscaling_model_path=request.upscaling_model_path,
-            upscaling_scale=request.upscaling_scale,
-            diffusers_kwargs=request.diffusers_kwargs,
-            profile=request.profile,
-            num_profiled_timesteps=request.num_profiled_timesteps,
-            profile_all_stages=request.profile_all_stages,
-            perf_dump_path=request.perf_dump_path,
-            output_path=request.output_path,
-            output_compression=request.output_compression,
-            output_quality=request.output_quality,
-            condition_inputs=chunk_inputs.condition_inputs,
-            realtime_chunk_size=chunk_size,
+            chunk_size=chunk_size,
         )
 
-    def prepare_next_request(
+    def _action_chunk_size(
         self,
         session: GenerateSession,
         server_args: ServerArgs,
         chunk: RealtimeChunkContext,
-    ) -> Req:
-        arch_config = server_args.pipeline_config.dit_config.arch_config
-        chunk_size = int(getattr(arch_config, "num_frames_per_block", 3))
+        chunk_size: int,
+    ) -> int:
         temporal_compression = int(
             server_args.pipeline_config.vae_config.arch_config.temporal_compression_ratio
         )
@@ -310,50 +248,17 @@ class SanaWMRealtimeAdapter(RealtimeModelAdapter):
                 action_chunk_size = (
                     segments[idx + 1] - segments[idx]
                 ) * temporal_compression
-        chunk_inputs = self._sample_chunk_inputs(session, chunk, action_chunk_size)
-        sampling_params = self._build_sampling_params(
-            session,
-            chunk,
-            chunk_inputs,
-            chunk_size,
-        )
-        batch = prepare_request(
-            server_args=server_args,
-            sampling_params=sampling_params,
-        )
-        batch.session = session.realtime_session
-        batch.realtime_session_id = session.id
-        batch.return_raw_frames = True
-        batch.block_idx = chunk.index
-        batch.realtime_event_id = self._state(session).latest_sampled_event_id
-        if session.request is not None:
-            # Forward the full transport config like the LingBot adapter does —
-            # the shared RawRGB output adapter / realtime_video_api consume
-            # preview width + pacing too; dropping them silently disabled both
-            # features for SANA-WM sessions.
-            batch.realtime_output_format = session.request.realtime_output_format
-            batch.realtime_preview_max_width = (
-                session.request.realtime_preview_max_width
-            )
-            batch.realtime_output_pacing = bool(session.request.realtime_output_pacing)
-        return batch
+        return action_chunk_size
 
-    async def send_output(
-        self,
-        ws: WebSocket,
-        session: GenerateSession,
-        result: OutputBatch,
-        batch: Req,
-    ) -> RealtimeFrameSendStats:
-        return await self.output_adapter.send(ws, session, result, batch)
+    def get_realtime_event_id(self, session: GenerateSession) -> int | None:
+        return self._state(session).latest_sampled_event_id
 
     def on_chunk_complete(self, session: GenerateSession, result: OutputBatch) -> None:
         if session.request is not None and self._raw_frame_count(result) == 0:
             session.request.max_chunks = session.generate_chunk_cnt + 1
         session.generate_chunk_completed()
 
-    def dispose(self, session: GenerateSession) -> None:
+    def clear_state(self, session: GenerateSession) -> None:
         state = session.adapter_state
         if isinstance(state, SanaWMRealtimeAdapterState):
             state.clear()
-        self.output_adapter.reset()
