+++ python/sglang/multimodal_gen/runtime/entrypoints/openai/realtime/generate_session.py
@@ -15,7 +15,7 @@ from sglang.multimodal_gen.runtime.realtime.session import (
 
 if TYPE_CHECKING:
     from sglang.multimodal_gen.runtime.entrypoints.openai.realtime.realtime_adapter import (
-        RealtimeModelAdapter,
+        BaseRealtimeModelAdapter,
     )
 
 
@@ -27,6 +27,8 @@ class RealtimeChunkContext:
 
 
 class GenerateSession:
+    """A realtime generation session"""
+
     def __init__(self):
         self.id = uuid4().hex
         self.request: RealtimeVideoGenerationsRequest | None = None
@@ -34,12 +36,12 @@ class GenerateSession:
         self.generate_chunk_cnt = 0
         self.current_chunk: RealtimeChunkContext | None = None
         self.realtime_session = RealtimeSession()
-        self.adapter: RealtimeModelAdapter | None = None
+        self.adapter: BaseRealtimeModelAdapter | None = None
         self.adapter_state: Any = None
         self.output_pace_next_send_at: float | None = None
         self.output_pace_last_event_id: int | None = None
 
-    def set_adapter(self, adapter: RealtimeModelAdapter):
+    def set_adapter(self, adapter: BaseRealtimeModelAdapter):
         self.adapter = adapter
         self.adapter_state = adapter.create_state()
