+++ python/sglang/multimodal_gen/runtime/pipelines_core/stages/causal_denoising.py
@@ -29,7 +29,10 @@ from sglang.multimodal_gen.runtime.platforms import (
     AttentionBackendEnum,
     current_platform,
 )
-from sglang.multimodal_gen.runtime.realtime.causal_state import RealtimeCausalDiTState
+from sglang.multimodal_gen.runtime.realtime.states import (
+    RealtimeCausalDiTState,
+    get_realtime_causal_dit_state,
+)
 from sglang.multimodal_gen.runtime.server_args import ServerArgs
 from sglang.multimodal_gen.runtime.utils.logging_utils import init_logger
 
@@ -314,7 +317,7 @@ class CausalDMDDenoisingStage(DenoisingStage):
         batch: Req,
     ) -> tuple[RealtimeCausalDiTState, bool]:
         if batch.session is not None:
-            state = batch.session.get_or_create_state(RealtimeCausalDiTState)
+            state = get_realtime_causal_dit_state(batch.session)
             return state, True
         return RealtimeCausalDiTState(), False
