--- python/sglang/multimodal_gen/runtime/realtime/causal_state.py
@@ -1,35 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-from sglang.multimodal_gen.runtime.realtime.session import BaseRealtimeState
-
-
-class RealtimeCausalDiTState(BaseRealtimeState):
-    """persist causal DiT cache and frame position across realtime chunks"""
-
-    def __init__(self):
-        super().__init__()
-        self.kv_cache = None
-        self.crossattn_cache = None
-        self.runtime_cache: dict = {}
-        self.current_chunk_start_frame: int = 0
-        self.chunk_idx: int = 0
-
-    def dispose(self) -> None:
-        self.kv_cache = None
-        self.crossattn_cache = None
-        self.runtime_cache.clear()
-        self.current_chunk_start_frame = 0
-        self.chunk_idx = 0
-
-
-class RealtimeCausalDecodeState(BaseRealtimeState):
-    """persist causal VAE decode cache and output frontier across chunks"""
-
-    def __init__(self):
-        super().__init__()
-        self.conv_cache: dict | None = None
-        self.next_dec_idx: int = 0
-
-    def dispose(self) -> None:
-        self.conv_cache = None
-        self.next_dec_idx = 0
