+++ python/sglang/multimodal_gen/test/unit/realtime/test_realtime_runtime.py
@@ -17,6 +17,7 @@ from sglang.multimodal_gen.runtime.entrypoints.openai.protocol import (
     RealtimeVideoGenerationsRequest,
 )
 from sglang.multimodal_gen.runtime.entrypoints.openai.realtime import (
+    realtime_adapter,
     realtime_video_api,
 )
 from sglang.multimodal_gen.runtime.entrypoints.openai.realtime.adapters import (
@@ -45,15 +46,15 @@ from sglang.multimodal_gen.runtime.pipelines_core.stages.realtime.input_validati
     RealtimeInputValidationStage,
     RealtimeInputValidationState,
 )
-from sglang.multimodal_gen.runtime.realtime.causal_state import (
-    RealtimeCausalDecodeState,
-)
-from sglang.multimodal_gen.runtime.realtime.condition_events import (
+from sglang.multimodal_gen.runtime.realtime.control_signals import (
     ControlStateTransition,
 )
 from sglang.multimodal_gen.runtime.realtime.session import (
     BaseRealtimeState,
     RealtimeSessionCache,
+)
+from sglang.multimodal_gen.runtime.realtime.states import (
+    RealtimeCausalDecodeState,
 )
 from sglang.multimodal_gen.runtime.utils.realtime_video import (
     RAW_RGB_CONTENT_TYPE,
@@ -153,11 +154,11 @@ def test_realtime_session_cache_rejects_missing_nonzero_chunk():
         raise AssertionError("expected missing realtime session to fail")
 
 
-def test_lingbot_realtime_state_uses_generic_condition_queue():
+def test_lingbot_realtime_state_uses_control_script_and_prompt_queues():
     state = lingbot_realtime.LingBotWorldRealtimeState()
 
     assert state.sample_camera_actions(3) == [[], [], []]
-    state.receive_camera_actions([["w"], ["a"], ["s"], ["d"]])
+    state.receive_camera_action_script([["w"], ["a"], ["s"], ["d"]])
     assert state.sample_camera_actions(3) == [["w"], ["a"], ["s"]]
     assert state.sample_camera_actions(3) == [["d"], [], []]
     assert state.sample_camera_actions(3) == [[], [], []]
@@ -166,6 +167,17 @@ def test_lingbot_realtime_state_uses_generic_condition_queue():
     assert state.has_prompt()
     assert state.sample_prompt() == "turn left"
     assert not state.has_prompt()
+
+
+def test_lingbot_realtime_camera_script_replaces_state_queue():
+    state = lingbot_realtime.LingBotWorldRealtimeState()
+
+    state.receive_camera_state(["w"], event_id=7)
+    state.receive_camera_action_script([["d"]], event_id=8)
+
+    assert state.sample_camera_actions(3) == [["d"], [], []]
+    assert state.latest_sampled_event_id == 8
+    assert state.sample_camera_actions(3) == [[], [], []]
 
 
 def test_lingbot_realtime_camera_events_preserve_short_presses():
@@ -210,7 +222,7 @@ def test_lingbot_realtime_camera_state_compacts_multiple_pending_updates():
 
 def test_sana_wm_realtime_camera_state_uses_sana_normalizer():
     state = sana_wm_realtime.SanaWMRealtimeAdapterState()
-    result = state.receive_camera_event_payload(
+    result = state.receive_camera_control_event_payload(
         {
             "mode": "state",
             "transitions": [
@@ -230,10 +242,10 @@ def test_sana_wm_realtime_adapter_preserves_requested_size():
     async def fake_save_image_to_path(image, target_path):
         return target_path
 
-    old_save_image_to_path = sana_wm_realtime.save_image_to_path
-    old_get_global_server_args = sana_wm_realtime.get_global_server_args
-    sana_wm_realtime.save_image_to_path = fake_save_image_to_path
-    sana_wm_realtime.get_global_server_args = lambda: SimpleNamespace(
+    old_save_image_to_path = realtime_adapter.save_image_to_path
+    old_get_global_server_args = realtime_adapter.get_global_server_args
+    realtime_adapter.save_image_to_path = fake_save_image_to_path
+    realtime_adapter.get_global_server_args = lambda: SimpleNamespace(
         input_save_path=None
     )
     try:
@@ -251,8 +263,8 @@ def test_sana_wm_realtime_adapter_preserves_requested_size():
 
         assert request.size == "832x480"
     finally:
-        sana_wm_realtime.save_image_to_path = old_save_image_to_path
-        sana_wm_realtime.get_global_server_args = old_get_global_server_args
+        realtime_adapter.save_image_to_path = old_save_image_to_path
+        realtime_adapter.get_global_server_args = old_get_global_server_args
 
 
 def test_lingbot_realtime_adapter_ingests_generic_events():
@@ -596,15 +608,16 @@ def test_lingbot_realtime_adapter_prepares_chunk_request(monkeypatch):
             prompt=sampling_params.prompt,
             condition_inputs=dict(sampling_params.condition_inputs),
             realtime_chunk_size=sampling_params.realtime_chunk_size,
+            session=None,
         )
 
     monkeypatch.setattr(
-        lingbot_realtime,
+        realtime_adapter,
         "build_sampling_params",
         fake_build_sampling_params,
     )
     monkeypatch.setattr(
-        lingbot_realtime,
+        realtime_adapter,
         "prepare_request",
         fake_prepare_backend_request,
     )
@@ -623,7 +636,7 @@ def test_lingbot_realtime_adapter_prepares_chunk_request(monkeypatch):
     assert batch.request_id == chunk.request_id
     assert batch.condition_inputs == {"camera_actions": [["w"], ["w"], ["w"]]}
     assert batch.realtime_chunk_size == 3
-    assert batch.session is session.realtime_session
+    assert batch.session is None
     assert batch.realtime_session_id == session.id
     assert batch.block_idx == 0
     assert batch.return_raw_frames is True
@@ -749,7 +762,7 @@ def test_lingbot_realtime_adapter_sends_stale_output_for_client_cutover():
     session = GenerateSession()
     session.set_adapter(adapter)
     state = adapter._state(session)
-    state.receive_camera_actions([["d"]], event_id=7)
+    state.receive_camera_action_script([["d"]], event_id=7)
     calls = []
 
     async def fake_send(ws, session_arg, result_arg, batch_arg):
